requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pinecone
google-generativeai
pytesseract
pdf2image
Pillow